refactor: move App and Modal to new layouts folder
